Move eval tests to separate file
